+++ src/components/CardList/StockCard/StockCard.Style.ts
