+++ src/pages/Search/Search.tsx
@@ -4,6 +4,7 @@ import { ResultInfo } from '@ts/Constants';
 import { RESULT_TYPE, STOCK_COUNTRY } from '@ts/Types';
 import { useIsMobile } from '@hooks/useIsMobile';
 import { useQuery } from '@hooks/useQuery';
+import StockCard from '@components/CardList/StockCard/StockCard';
 import { ContentsItemContainer, ContentsItemContent, ContentsItemTitle } from '@components/Common/ContentsItem.Style';
 import AntVoicePopUp from '@components/PopUp/AntiVoicePopUp/AntVoicePopUp';
 import ZipyoPopup from '@components/PopUp/ZipyoPopUp/ZipyoPopUp';
@@ -11,7 +12,6 @@ import SearchTitle from '@components/Search/SearchTitle/SearchTitle';
 import StockChart from '@components/Search/StockChart/StockChart';
 import StockWordCloud from '@components/Search/StockWordCloud/StockWordCloud';
 import SlideView from '@components/SlideView/SlideView';
-import StockCard from '@components/StockCard/StockCard';
 import ScoreSlotMachine from '@components/StockSlotMachine/StockSlotMachine';
 import { ScoreQuery, SearchSymbolNameQuery, StockRelevantQuery } from '@controllers/query';
 import AlertSVG from '@assets/alert.svg?react';
