+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/queryclient.ts","./src/vite-env.d.ts","./src/components/cardlist/cardlist.style.ts","./src/components/cardlist/cardlist.tsx","./src/components/cardlist/mobilestockcard/mobilestockcard.style.tsx","./src/components/cardlist/mobilestockcard/mobilestockcard.tsx","./src/components/cardlist/stockcard/stockcard.style.tsx","./src/components/cardlist/stockcard/stockcard.tsx","./src/components/common/common.props.ts","./src/components/common/common.type.ts","./src/components/common/common.tsx","./src/components/common/contentsitem.style.ts","./src/components/common/errorcomponent.tsx","./src/components/common/loadingcomponent.tsx","./src/components/home/indexscore/indexscore.style.ts","./src/components/home/indexscore/indexscore.tsx","./src/components/home/keywords/keywords.style.tsx","./src/components/home/keywords/keywords.tsx","./src/components/home/stocktable/stocktable.style.ts","./src/components/home/stocktable/stocktable.tsx","./src/components/pwausage/common.style.ts","./src/components/pwausage/android/android.style.tsx","./src/components/pwausage/android/android.tsx","./src/components/pwausage/ios/ios.style.ts","./src/components/pwausage/ios/ios.tsx","./src/components/popup/commonpopup.style.ts","./src/components/popup/commonpopup.tsx","./src/components/popup/antivoicepopup/antvoicepopup.style.ts","./src/components/popup/antivoicepopup/antvoicepopup.tsx","./src/components/popup/descentpopup/descentpopup.tsx","./src/components/popup/fearpopup/fearpopup.tsx","./src/components/popup/hotpopup/hotpopup.tsx","./src/components/popup/keywordpopup/keywordpopup.tsx","./src/components/popup/pwainfopopup/pwainfopopup.tsx","./src/components/popup/pwainfopopup/pwainfopopup.style.ts","./src/components/popup/risingpopup/risingpopup.tsx","./src/components/popup/zipyopopup/zipyopopup.style.ts","./src/components/popup/zipyopopup/zipyopopup.tsx","./src/components/search/searchtitle/searchtitle.style.ts","./src/components/search/searchtitle/searchtitle.tsx","./src/components/search/stockchart/stockchart.style.ts","./src/components/search/stockchart/stockchart.tsx","./src/components/search/stockwordcloud/stockwordcloud.style.ts","./src/components/search/stockwordcloud/stockwordcloud.tsx","./src/components/searchbar/searchbar.style.ts","./src/components/searchbar/searchbar.tsx","./src/components/slideview/slideview.style.ts","./src/components/slideview/slideview.tsx","./src/components/stockcard/stockcard.style.ts","./src/components/stockcard/stockcard.tsx","./src/components/stockslotmachine/stockslotmachine.tsx","./src/components/stockslotmachine/stockslotmachine.style.ts","./src/components/text/text.props.ts","./src/components/text/text.ts","./src/constants/stockscore.ts","./src/controllers/api.type.ts","./src/controllers/api.ts","./src/controllers/mock.ts","./src/controllers/query.ts","./src/hooks/usecanvas.ts","./src/hooks/useismobile.ts","./src/hooks/usequery.tsx","./src/hooks/useworker.ts","./src/layout/footer/footer.style.ts","./src/layout/footer/footer.tsx","./src/layout/header/header.style.ts","./src/layout/header/header.tsx","./src/layout/mainlayout/mainlayout.props.ts","./src/layout/mainlayout/mainlayout.style.ts","./src/layout/mainlayout/mainlayout.tsx","./src/pages/home/home.style.ts","./src/pages/home/home.tsx","./src/pages/search/search.style.ts","./src/pages/search/search.tsx","./src/pages/usage/usage.style.ts","./src/pages/usage/usage.tsx","./src/router/index.tsx","./src/styles/keyframes.ts","./src/styles/themes.ts","./src/ts/constants.ts","./src/ts/enums.ts","./src/ts/interfaces.ts","./src/ts/types.ts","./src/utils/canvas.ts","./src/utils/date.ts","./src/utils/delta.ts","./src/utils/detector.ts","./src/utils/localstorage.ts","./src/utils/scoreconvert.ts","./src/utils/chunkarray.ts","./src/utils/getframerate.ts","./src/utils/wasm/wordcloudwasm.ts","./src/utils/wasm/wordcloud.wasm.d.ts","./src/utils/worker/generatewordcloud.ts"],"version":"5.7.2"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/queryclient.ts","./src/vite-env.d.ts","./src/components/cardlist/cardlist.tsx","./src/components/cardlist/stockcard/stockcard.style.ts","./src/components/cardlist/stockcard/stockcard.tsx","./src/components/common/common.props.ts","./src/components/common/common.type.ts","./src/components/common/common.tsx","./src/components/common/contentsitem.style.ts","./src/components/common/errorcomponent.tsx","./src/components/common/loadingcomponent.tsx","./src/components/home/indexscore/indexscore.style.ts","./src/components/home/indexscore/indexscore.tsx","./src/components/home/keywords/keywords.style.tsx","./src/components/home/keywords/keywords.tsx","./src/components/home/stocktable/stocktable.style.ts","./src/components/home/stocktable/stocktable.tsx","./src/components/pwausage/common.style.ts","./src/components/pwausage/android/android.style.tsx","./src/components/pwausage/android/android.tsx","./src/components/pwausage/ios/ios.style.ts","./src/components/pwausage/ios/ios.tsx","./src/components/popup/commonpopup.style.ts","./src/components/popup/commonpopup.tsx","./src/components/popup/antivoicepopup/antvoicepopup.style.ts","./src/components/popup/antivoicepopup/antvoicepopup.tsx","./src/components/popup/descentpopup/descentpopup.tsx","./src/components/popup/fearpopup/fearpopup.tsx","./src/components/popup/hotpopup/hotpopup.tsx","./src/components/popup/keywordpopup/keywordpopup.tsx","./src/components/popup/pwainfopopup/pwainfopopup.tsx","./src/components/popup/pwainfopopup/pwainfopopup.style.ts","./src/components/popup/risingpopup/risingpopup.tsx","./src/components/popup/zipyopopup/zipyopopup.style.ts","./src/components/popup/zipyopopup/zipyopopup.tsx","./src/components/search/searchtitle/searchtitle.style.ts","./src/components/search/searchtitle/searchtitle.tsx","./src/components/search/stockchart/stockchart.style.ts","./src/components/search/stockchart/stockchart.tsx","./src/components/search/stockwordcloud/stockwordcloud.style.ts","./src/components/search/stockwordcloud/stockwordcloud.tsx","./src/components/searchbar/searchbar.style.ts","./src/components/searchbar/searchbar.tsx","./src/components/slideview/slideview.style.ts","./src/components/slideview/slideview.tsx","./src/components/stockslotmachine/stockslotmachine.tsx","./src/components/stockslotmachine/stockslotmachine.style.ts","./src/components/text/text.props.ts","./src/components/text/text.ts","./src/constants/stockscore.ts","./src/controllers/api.type.ts","./src/controllers/api.ts","./src/controllers/mock.ts","./src/controllers/query.ts","./src/hooks/usecanvas.ts","./src/hooks/useismobile.ts","./src/hooks/usequery.tsx","./src/hooks/useworker.ts","./src/layout/footer/footer.style.ts","./src/layout/footer/footer.tsx","./src/layout/header/header.style.ts","./src/layout/header/header.tsx","./src/layout/mainlayout/mainlayout.props.ts","./src/layout/mainlayout/mainlayout.style.ts","./src/layout/mainlayout/mainlayout.tsx","./src/pages/home/home.style.ts","./src/pages/home/home.tsx","./src/pages/search/search.style.ts","./src/pages/search/search.tsx","./src/pages/usage/usage.style.ts","./src/pages/usage/usage.tsx","./src/router/index.tsx","./src/styles/keyframes.ts","./src/styles/themes.ts","./src/ts/constants.ts","./src/ts/enums.ts","./src/ts/interfaces.ts","./src/ts/types.ts","./src/utils/canvas.ts","./src/utils/date.ts","./src/utils/delta.ts","./src/utils/detector.ts","./src/utils/localstorage.ts","./src/utils/scoreconvert.ts","./src/utils/chunkarray.ts","./src/utils/getframerate.ts","./src/utils/wasm/wordcloudwasm.ts","./src/utils/wasm/wordcloud.wasm.d.ts","./src/utils/worker/generatewordcloud.ts"],"version":"5.7.2"}
